% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cdm.R
\name{check_qmatrix_identification}
\alias{check_qmatrix_identification}
\title{Check Q-matrix Identification (cdmTools)}
\usage{
check_qmatrix_identification(q_matrix, model_name)
}
\description{
Check Q-matrix Identification (cdmTools)
}
